Check for commonly misspelled words during the CI build
Use codespell to check for commonly misspelled words during the Travis CI build. In the event of false positives, the words can be added to extras/codespell-ignore-words-list.txt, after which codespell will ignore those words.

I excluded src/minmea from the check because this appears to be an external resource.